refactor: s3관련 로직 분리
